Move arch context layouts into arch crate
