+++ kernel/src/arch/riscv/mod.rs
@@ -19,7 +19,10 @@ use core::{
     cell::Cell,
     sync::atomic::{compiler_fence, Ordering},
 };
-pub use bluekernel_arch::riscv::{irq, trap_entry};
+pub use bluekernel_arch::riscv::{
+    context::{Context, IsrContext},
+    irq, trap_entry,
+};
 pub(super) use trap::{handle_ecall_switch_from_raw, might_switch_context_from_raw};
 
 pub(crate) const NR_SWITCH: usize = !0;
@@ -363,98 +366,6 @@ pub(crate) extern "C" fn switch_context_with_hook(hook: *mut ContextSwitchHookHo
 pub(crate) extern "C" fn restore_context_with_hook(hook: *mut ContextSwitchHookHolder) -> ! {
     switch_context_with_hook(hook);
     unreachable!("Should have switched to another thread");
-}
-
-// This context is used when we are performing context switching in
-// thread mode or in the first level ISR.
-#[cfg_attr(target_pointer_width = "64", repr(C, align(16)))]
-#[cfg_attr(target_pointer_width = "32", repr(C, align(8)))]
-#[derive(Default, Debug)]
-pub(crate) struct Context {
-    pub ra: usize,
-    pub mepc: usize,
-    pub gp: usize,
-    pub tp: usize,
-    pub t0: usize,
-    pub t1: usize,
-    pub t2: usize,
-    pub fp: usize,
-    pub a0: usize,
-    pub a1: usize,
-    pub a2: usize,
-    pub a3: usize,
-    pub a4: usize,
-    pub a5: usize,
-    pub a6: usize,
-    pub a7: usize,
-    pub t3: usize,
-    pub t4: usize,
-    pub t5: usize,
-    pub t6: usize,
-    pub s1: usize,
-    pub s2: usize,
-    pub s3: usize,
-    pub s4: usize,
-    pub s5: usize,
-    pub s6: usize,
-    pub s7: usize,
-    pub s8: usize,
-    pub s9: usize,
-    pub s10: usize,
-    pub s11: usize,
-    // So that it's 16-byte aligned.
-    pub padding: usize,
-}
-
-#[repr(C, align(16))]
-#[derive(Default, Debug)]
-pub(crate) struct IsrContext {
-    pub mstatus: usize,
-    pub mcause: usize,
-    pub mtval: usize,
-    pub mepc: usize,
-}
-
-impl Context {
-    #[inline]
-    pub(crate) fn init(&mut self) -> &mut Self {
-        self.gp = Self::__global_pointer();
-        self
-    }
-
-    #[inline]
-    pub(crate) fn __global_pointer() -> usize {
-        let gp_val: usize;
-        unsafe {
-            core::arch::asm!("la {}, __global_pointer$", out(reg) gp_val,
-                             options(nostack, nomem))
-        }
-        gp_val
-    }
-
-    // We are following C-ABI, since Rust ABI is not stabilized.
-    // FIXME: rustc miscompiles it if inlined.
-    #[inline(never)]
-    pub(crate) fn set_return_address(&mut self, pc: usize) -> &mut Self {
-        self.mepc = pc;
-        self
-    }
-
-    #[inline(never)]
-    pub(crate) fn set_arg(&mut self, index: usize, val: usize) -> &mut Self {
-        match index {
-            0 => self.a0 = val,
-            1 => self.a1 = val,
-            2 => self.a2 = val,
-            3 => self.a3 = val,
-            4 => self.a4 = val,
-            5 => self.a5 = val,
-            6 => self.a6 = val,
-            7 => self.a7 = val,
-            _ => {}
-        }
-        self
-    }
 }
 
 pub(crate) extern "C" fn bootstrap() {
